spring/springmvc1/item-service/src/main/java/hello/itemservice/domain/item/ItemRepository.java
package hello.itemservice.domain.item;

import java.util.ArrayList;
import java.util.Collection;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import org.springframework.stereotype.Repository;

@Repository
public class ItemRepository {
    private static final Map<Long, Item> store = new HashMap<>();
    private static long sequence = 0L;

    public Item save(Item item) {
        item.setId(++sequence);
        store.put(item.getId(), item);

        return item;
    }

    public Item findById(Long id){
        return store.get(id);
    }

    public List<Item> findAll() {
        return new ArrayList<>(store.values());
    }

    // ! 원래 실제로는 updateParam의 타입으로 itemParamDto 같은 클래스를 만들어서 쓰는게 좋음. id 필드를 사용하고 있지 않기 때문
    public void update(Long itemId, Item updateParam) { 
        Item findItem = findById(itemId);

        findItem.setItemName(updateParam.getItemName());
        findItem.setPrice(updateParam.getPrice());
        findItem.setQuantity(updateParam.getQuantity());
    }

    public void clearStore() {
        store.clear();
    }
} 
